# Conflicts:
#	translator/test_infra/tests/testssol/env/revm.rs

+++ translator/test_infra/sol/a_plus_b.sol
@@ -5,15 +5,15 @@
 pragma solidity ^0.8.10;
 
 contract APlusB {
-    function plus() public pure returns (uint) {
+    function plus() public pure returns (uint128) {
         return plus_1(13, 14);
     }
 
-    function plus_1(uint a, uint b) internal pure returns (uint) {
+    function plus_1(uint128 a, uint128 b) internal pure returns (uint128) {
         return a + b;
     }
 
-    function minus() public pure returns (uint) {
+    function minus() public pure returns (uint128) {
         return 14 - 13;
     }
 }
