+++ translator/test_infra/sol/loop/ignore_for.sol
@@ -4,9 +4,9 @@ pragma solidity ^0.8.0;
 
 library WhileLoop {
     // # sum (2,10) 20
-    function sum(uint inc, uint limit) public pure returns (uint) {
-        uint result = 0;
-        for(uint i=0; i<limit; i++) {
+    function sum(uint128 inc, uint128 limit) public pure returns (uint128) {
+        uint128 result = 0;
+        for(uint128 i=0; i<limit; i++) {
             result += inc;
         }
         return result;
