+++ translator/test_infra/sol/loop/ignore_while.sol
@@ -4,8 +4,8 @@ pragma solidity ^0.8.0;
 
 library WhileLoop {
     // # sum (2,10) 20
-    function sum(uint inc, uint limit) public pure returns (uint) {
-        uint result = 0;
+    function sum(uint128 inc, uint128 limit) public pure returns (uint128) {
+        uint128 result = 0;
         while (limit > 0) {
             --limit;
             result += inc;
