+++ translator/test_infra/sol/uniswap/4_contract_param_ignore.sol
@@ -5,10 +5,10 @@
 pragma solidity >=0.8.10;
 
 contract TestParam {
-    uint public totalSupply;
+    uint128 public totalSupply;
 
     // # get_set(10) 10
-    function get_set(uint value) public returns(uint){
+    function get_set(uint128 value) public returns(uint128){
         totalSupply = value;
 
         return totalSupply;
