+++ translator/test_infra/sol/uniswap/7_constructor_ignore.sol
@@ -9,7 +9,7 @@ contract TestConstruct {
     bytes32 public DOMAIN_SEPARATOR;
 
     constructor() {
-        uint chainId = 0;
+        uint128 chainId = 0;
         DOMAIN_SEPARATOR = keccak256(
             abi.encode(
                 keccak256('EIP712Domain(string name,string version,uint256 chainId,address verifyingContract)'),
