+++ translator/test_infra/sol/uniswap/8_mapping_ignore.sol
@@ -7,10 +7,10 @@ pragma solidity >=0.8.10;
 import "./l_SafeMath.sol";
 
 contract TestMapping {
-    mapping(address => uint) public balanceOf;
+    mapping(address => uint128) public balanceOf;
 
     // # mint(0x1,10) 10
-    function mint(address to, uint value) public returns(uint){
+    function mint(address to, uint128 value) public returns(uint128){
         balanceOf[to] = balanceOf[to] + value;
         return balanceOf[to];
     }
